# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Gym CRM REST API
  version: 1.0.0
  description: |
    REST API for the Gym CRM system.

    Public endpoints are limited to trainee registration, trainer registration,
    and login. Every other endpoint requires a Bearer JWT.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/login:
    post:
      security: []
      tags:
        - Authentication
      summary: Authenticate a user and issue a JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials or user temporarily blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/logout:
    post:
      tags:
        - Authentication
      summary: Log out and revoke the current JWT
      operationId: logout
      responses:
        '200':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/password:
    put:
      tags:
        - Authentication
      summary: Change the authenticated user's password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid JWT, or incorrect old password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, expired, or revoked bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: John.Smith
        password:
          type: string
          format: password
          writeOnly: true
          example: pass123456
    TokenResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
      properties:
        accessToken:
          type: string
          description: Signed JWT access token
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          description: Token lifetime in seconds
          example: 1800
    ChangePasswordRequest:
      type: object
      required:
        - oldPassword
        - newPassword
      properties:
        oldPassword:
          type: string
          format: password
          writeOnly: true
        newPassword:
          type: string
          format: password
          writeOnly: true
...
    ErrorResponse:
      type: object
      required:
        - status
        - error
        - message
      properties:
        status:
          type: integer
          format: int32
          example: 401
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Missing or invalid bearer token